refactor: rename `run` to `runner`
