+++ src/cli.ts
@@ -3,7 +3,7 @@
 import { readFileSync } from 'fs';
 import { resolve } from 'path';
 import * as yargs from 'yargs';
-import { run, RunOptions } from './run';
+import { run, RunnerOptions } from './runner';
 
 const pkgJson = JSON.parse(readFileSync(resolve(__dirname, '../package.json'), 'utf-8'));
 
@@ -60,7 +60,7 @@ const yargsInstance = yargs
  * Parse command line arguments.
  * @returns Runner options.
  */
-export function parseArgv(argv: string[]): RunOptions {
+export function parseArgv(argv: string[]): RunnerOptions {
   const parsedArgv = yargsInstance.parseSync(argv.slice(2));
   const patterns: string[] = parsedArgv._.map((pattern) => pattern.toString());
   return {
